pyproject.toml
[tool.poetry]
name = "venus"
version = "0.1.0"
description = ""
authors = []

packages = [
    { include = "venus", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
pydantic = "^1.10.2"
requests = "^2.28.1"
auth0-python = "^3.23.1"
dataclasses-json = "^0.5.7"
types-requests = "^2.28.11"
PyJWT = "^2.6.0"
posthog = "^2.3.0"

[tool.poetry.scripts]
start = "venus.main:start"


[tool.poetry.dev-dependencies]
mypy = "^0.971"
pytest = "^7.1.3"
black = "^22.8.0"
flake8 = "^5.0.4"
pre-commit = "^2.20.0"
pytest-env = "^0.6.2"
pytest-docker = "^1.0.1"
fern-fern-nursery = {version = "0.0.191", source = "fern-dev"}

[[tool.poetry.source]]
name = "fern-dev"
url = "https://pypi-dev.buildwithfern.com/"
default = false
secondary = true

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
line_length = 79
force_single_line = true
combine_as_imports = true
lines_between_types = 1
lines_after_imports = 2
src_paths = ["src", "tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
